+++ XuanYu.Editor.UI/Shell/BottomDockHost.axaml.cs
@@ -1,8 +1,0 @@
-using Avalonia.Controls;
-
-namespace XuanYu.Editor.UI;
-
-public partial class BottomDockHost : UserControl
-{
-    public BottomDockHost() => InitializeComponent();
-}